Create LICENSE
In this commit I am adding a license to my project. I chose the MIT License as it is the license that best applies to the general public. Short and simple permissive license that allows others to use and distribute my code. However, it also does not hold me liable for others using my code with malicious intent.